Time: 81 ms (60.43%) | Memory: 79 MB (93.36%) - LeetSync

+++ 2626-count-the-number-of-good-subarrays/count-the-number-of-good-subarrays.cpp
@@ -3,7 +3,7 @@ public:
 
     long long countGood(vector<int>& nums, int k) {
         long long ans = 0, l = 0, r = l + 1, n = nums.size(), cnt = 0;
-        unordered_map<int, long long> mp;
+        unordered_map<int, int> mp;
         mp[nums[l]]++;
         while(r < n){
             cnt += mp[nums[r]];
